+++ .ort-data/curations-dir/voting-booth.yml
@@ -1,7 +1,7 @@
-- id: "Yarn::agora-gui-booth"
+- id: "Yarn::voting-booth"
   curations:
     comment: "This package needs to be downloaded from git"
     vcs:
       type: "git"
-      url: "https://github.com/agoravoting/agora-gui-common.git"
+      url: "https://github.com/sequentech/common-ui.git"
       revision: "4.0.1"
