+++ .ort-data/disclosure_document.ftl
@@ -1,7 +1,7 @@
 [#--
 Copyright (C) 2020 HERE Europe B.V.
 Copyright (C) 2020-2021 Bosch.IO GmbH
-Copyright (C) 2021 Agora Voting SL
+Copyright (C) 2021 Sequent Tech Inc
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
