+++ AUTHORS
@@ -1,4 +1,4 @@
-Eduardo Robles Elvira <edulix AT agoravoting DOT com>
-David Ruescas <david AT agoravoting DOT com>
+Eduardo Robles Elvira <edulix@sequentech.io>
+David Ruescas <david@sequentech.io>
 Daniel García Moreno <danigm AT wadobo DOT com>
-Félix Robles Elvira <felix AT agoravoting DOT com>
+Félix Robles Elvira <felix@sequentech.io>
