+++ README.md
@@ -1,9 +1,9 @@
-# agora-tally [![tests_badge]][tests_link]
+# tally-methods [![tests_badge]][tests_link]
 
-`agora-tally` provides functions for tallying election questions.
+`tally-methods` provides functions for tallying election questions.
 
-[tests_badge]: https://github.com/agoravoting/agora-tally/workflows/Test%20python/badge.svg
-[tests_link]: https://github.com/agoravoting/agora-tally/actions?query=workflow%3A%22Test+python%22
+[tests_badge]: https://github.com/sequentech/tally-methods/workflows/Test%20python/badge.svg
+[tests_link]: https://github.com/sequentech/tally-methods/actions?query=workflow%3A%22Test+python%22
 
 ### Entry functions
 
@@ -66,7 +66,7 @@ test/testing_setup.sh
 
 is provided to do this. When testing you have to activate the environment with
 
-```workon agora-tally```
+```workon tally-methods```
 
 Then run the tests with
 
@@ -74,21 +74,21 @@ Then run the tests with
 
 # License
 
-Copyright (C) 2015 Agora Voting SL and/or its subsidiary(-ies).
-Contact: legal@agoravoting.com
+Copyright (C) 2015 Sequent Tech Inc and/or its subsidiary(-ies).
+Contact: legal@sequentech.io
 
-This file is part of the agora-core-view module of the Agora Voting project.
+This file is part of the sequent-core-view module of the Sequent Tech project.
 
 This program is distributed in the hope that it will be useful, but WITHOUT ANY
 WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A
 PARTICULAR PURPOSE.  See the GNU General Public License for more details.
 
 Commercial License Usage
-Licensees holding valid commercial Agora Voting project licenses may use this
+Licensees holding valid commercial Sequent Tech project licenses may use this
 file in accordance with the commercial license agreement provided with the
 Software or, alternatively, in accordance with the terms contained in
-a written agreement between you and Agora Voting SL. For licensing terms and
-conditions and further information contact us at legal@agoravoting.com .
+a written agreement between you and Sequent Tech Inc. For licensing terms and
+conditions and further information contact us at legal@sequentech.io .
 
 GNU Affero General Public License Usage
 Alternatively, this file may be used under the terms of the GNU Affero General
