+++ tally_methods/tally.py
@@ -1,22 +1,22 @@
 #!/usr/bin/env python
 
-# This file is part of agora-tally.
-# Copyright (C) 2013-2016  Agora Voting SL <agora@agoravoting.com>
+# This file is part of tally-methods.
+# Copyright (C) 2013-2016  Sequent Tech Inc <legal@sequentech.io>
 
-# agora-tally is free software: you can redistribute it and/or modify
+# tally-methods is free software: you can redistribute it and/or modify
 # it under the terms of the GNU Affero General Public License as published by
 # the Free Software Foundation, either version 3 of the License.
 
-# agora-tally  is distributed in the hope that it will be useful,
+# tally-methods  is distributed in the hope that it will be useful,
 # but WITHOUT ANY WARRANTY; without even the implied warranty of
 # MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 # GNU Affero General Public License for more details.
 
 # You should have received a copy of the GNU Affero General Public License
-# along with agora-tally.  If not, see <http://www.gnu.org/licenses/>.
+# along with tally-methods.  If not, see <http://www.gnu.org/licenses/>.
 
 
-from agora_tally.voting_systems.base import get_voting_system_by_id, BlankVoteException
+from tally_methods.voting_systems.base import get_voting_system_by_id, BlankVoteException
 
 import copy
 import glob
