+++ tally_methods/voting_systems/borda_custom.py
@@ -1,33 +1,30 @@
-# This file is part of agora-tally.
-# Copyright (C) 2013-2021  Agora Voting SL <agora@agoravoting.com>
+# This file is part of tally-methods.
+# Copyright (C) 2013-2021  Sequent Tech Inc <legal@sequentech.io>
 
-# agora-tally is free software: you can redistribute it and/or modify
+# tally-methods is free software: you can redistribute it and/or modify
 # it under the terms of the GNU Affero General Public License as published by
 # the Free Software Foundation, either version 3 of the License.
 
-# agora-tally  is distributed in the hope that it will be useful,
+# tally-methods  is distributed in the hope that it will be useful,
 # but WITHOUT ANY WARRANTY; without even the implied warranty of
 # MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 # GNU Affero General Public License for more details.
 
 # You should have received a copy of the GNU Affero General Public License
-# along with agora-tally.  If not, see <http://www.gnu.org/licenses/>.
-
-import copy
-from collections import defaultdict
+# along with tally-methods.  If not, see <http://www.gnu.org/licenses/>.
 
 from .base import (
     BaseVotingSystem, 
     BaseTally, 
-    BlankVoteException,
-    get_key,
-    WeightedChoice
+    WeightedChoice, 
+    get_key
 )
+from .borda import BordaTally
 
-class Cumulative(BaseVotingSystem):
+class BordaCustom(BaseVotingSystem):
     '''
-    Defines the helper functions that allows agora to manage a
-    cumulative voting system.
+    Defines the helper functions that allows sequent to manage an OpenSTV-based
+    Nauru Borda voting system.
     '''
 
     @staticmethod
@@ -36,29 +33,28 @@ class Cumulative(BaseVotingSystem):
         Returns the identifier of the voting system, used internally to
         discriminate  the voting system used in an election
         '''
-        return 'cumulative'
+        return 'borda-custom'
 
     @staticmethod
     def get_description():
-        return _('Cumulative voting')
+        return _('Custom Borda Count voting')
 
     @staticmethod
     def create_tally(question, question_num):
         '''
         Create object that helps to compute the tally
         '''
-        return CumulativeTally(
+        return BordaCustomTally(
             question=question,
             question_num=question_num
         )
 
-class CumulativeTally(BaseTally):
-    '''
-    Class used to tally a cumulative election
-    '''
+class BordaCustomTally(BordaTally):
     def init(self):
         def custom_subparser(decoded_ballot, question, withdrawals):
             answers = set()
+
+            weights = question['borda_custom_weights']
 
             for answer in decoded_ballot["answers"]:
                 if answer['selected'] < 0 or answer['id'] in withdrawals:
@@ -67,10 +63,30 @@ class CumulativeTally(BaseTally):
                 answers.add(
                     WeightedChoice(
                         key=get_key(answer),
-                        points=(answer['selected'] + 1),
+                        points=weights[answer['selected']],
                         answer_id=answer['id']
                     )
                 )
+            
+            # Check for invalid votes:
+            selection = [
+                answer['selected']
+                for answer in decoded_ballot["answers"]
+                if answer['selected'] >= 0
+            ]
+            # - no position is repeated
+            if len(selection) != len(set(selection)):
+                raise Exception()
+
+            selection_sorted = sorted(selection)
+            should_be_selection_sorted = [
+                index
+                for index, _ in enumerate(selection_sorted)
+            ]
+            # - no missing position in-between
+            if selection_sorted != should_be_selection_sorted:
+                raise Exception()
+
             return frozenset(answers)
 
         self.custom_subparser = custom_subparser
