+++ tally_methods/voting_systems/borda_nauru.py
@@ -1,17 +1,17 @@
-# This file is part of agora-tally.
-# Copyright (C) 2013-2021  Agora Voting SL <agora@agoravoting.com>
+# This file is part of tally-methods.
+# Copyright (C) 2013-2021  Sequent Tech Inc <legal@sequentech.io>
 
-# agora-tally is free software: you can redistribute it and/or modify
+# tally-methods is free software: you can redistribute it and/or modify
 # it under the terms of the GNU Affero General Public License as published by
 # the Free Software Foundation, either version 3 of the License.
 
-# agora-tally  is distributed in the hope that it will be useful,
+# tally-methods  is distributed in the hope that it will be useful,
 # but WITHOUT ANY WARRANTY; without even the implied warranty of
 # MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 # GNU Affero General Public License for more details.
 
 # You should have received a copy of the GNU Affero General Public License
-# along with agora-tally.  If not, see <http://www.gnu.org/licenses/>.
+# along with tally-methods.  If not, see <http://www.gnu.org/licenses/>.
 
 from .base import (
     BaseVotingSystem, 
@@ -23,7 +23,7 @@ from .borda import BordaTally
 
 class BordaNauru(BaseVotingSystem):
     '''
-    Defines the helper functions that allows agora to manage an OpenSTV-based
+    Defines the helper functions that allows sequent to manage an OpenSTV-based
     Nauru Borda voting system.
     '''
