+++ tally_methods/voting_systems/plurality_at_large.py
@@ -1,17 +1,17 @@
-# This file is part of agora-tally.
-# Copyright (C) 2013-2021  Agora Voting SL <agora@agoravoting.com>
+# This file is part of tally-methods.
+# Copyright (C) 2013-2021  Sequent Tech Inc <legal@sequentech.io>
 
-# agora-tally is free software: you can redistribute it and/or modify
+# tally-methods is free software: you can redistribute it and/or modify
 # it under the terms of the GNU Affero General Public License as published by
 # the Free Software Foundation, either version 3 of the License.
 
-# agora-tally  is distributed in the hope that it will be useful,
+# tally-methods  is distributed in the hope that it will be useful,
 # but WITHOUT ANY WARRANTY; without even the implied warranty of
 # MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 # GNU Affero General Public License for more details.
 
 # You should have received a copy of the GNU Affero General Public License
-# along with agora-tally.  If not, see <http://www.gnu.org/licenses/>.
+# along with tally-methods.  If not, see <http://www.gnu.org/licenses/>.
 
 from .base import (
     BaseVotingSystem, 
@@ -19,12 +19,11 @@ from .base import (
     WeightedChoice, 
     get_key
 )
-from .borda import BordaTally
 
-class BordaCustom(BaseVotingSystem):
+class PluralityAtLarge(BaseVotingSystem):
     '''
-    Defines the helper functions that allows agora to manage an OpenSTV-based
-    Nauru Borda voting system.
+    Defines the helper functions that allows sequent to manage an OpenSTV-based
+    Plurality at large voting system.
     '''
 
     @staticmethod
@@ -33,28 +32,29 @@ class BordaCustom(BaseVotingSystem):
         Returns the identifier of the voting system, used internally to
         discriminate  the voting system used in an election
         '''
-        return 'borda-custom'
+        return 'plurality-at-large'
 
     @staticmethod
     def get_description():
-        return _('Custom Borda Count voting')
+        return _('Plurality at large voting')
 
     @staticmethod
     def create_tally(question, question_num):
         '''
         Create object that helps to compute the tally
         '''
-        return BordaCustomTally(
+        return PluralityAtLargeTally(
             question=question,
             question_num=question_num
         )
 
-class BordaCustomTally(BordaTally):
+class PluralityAtLargeTally(BaseTally):
+    '''
+    Class used to tally an election
+    '''
     def init(self):
         def custom_subparser(decoded_ballot, question, withdrawals):
             answers = set()
-
-            weights = question['borda_custom_weights']
 
             for answer in decoded_ballot["answers"]:
                 if answer['selected'] < 0 or answer['id'] in withdrawals:
@@ -63,30 +63,11 @@ class BordaCustomTally(BordaTally):
                 answers.add(
                     WeightedChoice(
                         key=get_key(answer),
-                        points=weights[answer['selected']],
-                        answer_id=answer['id']
+                        answer_id=answer['id'],
+                        points=(answer['selected'] + 1)
                     )
                 )
-            
-            # Check for invalid votes:
-            selection = [
-                answer['selected']
-                for answer in decoded_ballot["answers"]
-                if answer['selected'] >= 0
-            ]
-            # - no position is repeated
-            if len(selection) != len(set(selection)):
-                raise Exception()
-
-            selection_sorted = sorted(selection)
-            should_be_selection_sorted = [
-                index
-                for index, _ in enumerate(selection_sorted)
-            ]
-            # - no missing position in-between
-            if selection_sorted != should_be_selection_sorted:
-                raise Exception()
-
             return frozenset(answers)
 
         self.custom_subparser = custom_subparser
+
