+++ test/desborda_test.py
@@ -1,20 +1,20 @@
 #!/usr/bin/env python3
 # -*- coding:utf-8 -*-
 
-# This file is part of agora-tally.
-# Copyright (C) 2017-2021  Agora Voting SL <agora@agoravoting.com>
+# This file is part of tally-methods.
+# Copyright (C) 2017-2021  Sequent Tech Inc <legal@sequentech.io>
 
-# agora-tally is free software: you can redistribute it and/or modify
+# tally-methods is free software: you can redistribute it and/or modify
 # it under the terms of the GNU Affero General Public License as published by
 # the Free Software Foundation, either version 3 of the License.
 
-# agora-tally  is distributed in the hope that it will be useful,
+# tally-methods  is distributed in the hope that it will be useful,
 # but WITHOUT ANY WARRANTY; without even the implied warranty of
 # MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 # GNU Affero General Public License for more details.
 
 # You should have received a copy of the GNU Affero General Public License
-# along with agora-tally.  If not, see <http://www.gnu.org/licenses/>.
+# along with tally-methods.  If not, see <http://www.gnu.org/licenses/>.
 
 import string
 import re
@@ -24,8 +24,8 @@ import os
 from os import urandom
 import tempfile
 
-from agora_tally.ballot_codec.nvotes_codec import NVotesCodec
-from agora_tally import file_helpers
+from tally_methods.ballot_codec.sequent_codec import NVotesCodec
+from tally_methods import file_helpers
 
 def remove_spaces(in_str):
     return re.sub(r"[ \t\r\f\v]*", "", in_str)
