+++ test/testing_setup.sh
@@ -2,9 +2,9 @@
 
 sudo apt-get install -y git virtualenvwrapper
 source /etc/bash_completion.d/virtualenvwrapper
-if [ ! -d $HOME/.virtualenvs/agora-tally ]
+if [ ! -d $HOME/.virtualenvs/tally-methods ]
 then
-    mkvirtualenv agora-tally
+    mkvirtualenv tally-methods
 fi
-workon agora-tally
+workon tally-methods
 #pip install -r testing_requirements.txt
